// src/generated/models/Ava_bbu_beurteilungbesondererbedurfnissesModel.ts
export const Ava_bbu_beurteilungbesondererbedurfnissesava_bbu_status = {
  100000000: 'offen',
  100000001: 'laufend',
  100000002: 'abgeschlossen',
  100000003: 'abgelehnt',
  669130001: 'nichtwahrgenommen'
} as const;
export type Ava_bbu_beurteilungbesondererbedurfnissesava_bbu_status = keyof typeof Ava_bbu_beurteilungbesondererbedurfnissesava_bbu_status;
export const Ava_bbu_beurteilungbesondererbedurfnissesava_bbu_zustimmung = {
  0: 'Nein',
  1: 'Ja'
} as const;
export type Ava_bbu_beurteilungbesondererbedurfnissesava_bbu_zustimmung = keyof typeof Ava_bbu_beurteilungbesondererbedurfnissesava_bbu_zustimmung;
export const Ava_bbu_beurteilungbesondererbedurfnissesava_gesprachstyp = {
  751960000: 'Erstgespr_ch',
  751960001: 'Folgegespr_ch'
} as const;
export type Ava_bbu_beurteilungbesondererbedurfnissesava_gesprachstyp = keyof typeof Ava_bbu_beurteilungbesondererbedurfnissesava_gesprachstyp;
export const Ava_bbu_beurteilungbesondererbedurfnissesstatecode = {
  0: 'Aktiv',
  1: 'Inaktiv'
} as const;
export type Ava_bbu_beurteilungbesondererbedurfnissesstatecode = keyof typeof Ava_bbu_beurteilungbesondererbedurfnissesstatecode;
export const Ava_bbu_beurteilungbesondererbedurfnissesstatuscode = {
  1: 'Aktiv',
  2: 'Inaktiv'
} as const;
export type Ava_bbu_beurteilungbesondererbedurfnissesstatuscode = keyof typeof Ava_bbu_beurteilungbesondererbedurfnissesstatuscode;

export interface Ava_bbu_beurteilungbesondererbedurfnissesBase {
  ava_abgeschlossen_am?: string;
  ava_bbu_allgemeinebeurteilungdurchgefuhrtam?: string;
  ava_bbu_beurteilungbesondererbedurfnisseid: string;
  "ava_bbu_Gesprachsteilnehmerin@odata.bind"?: string;
  "ava_bbu_Klient@odata.bind": string;
  ava_bbu_medbeurteilungdurchgefuhrtam?: string;
  ava_bbu_medbeurteilungnotwendig?: string;
  ava_bbu_psycheinschatzungdurchgefuhrt?: string;
  ava_bbu_psycheinschatzungsbedarffestgestell?: string;
  "ava_bbu_Standort@odata.bind": string;
  ava_bbu_status?: Ava_bbu_beurteilungbesondererbedurfnissesava_bbu_status;
  ava_bbu_ubermittlunganbfa?: string;
  ava_bbu_ubermittlunganklientin?: string;
  ava_bbu_zustimmung?: Ava_bbu_beurteilungbesondererbedurfnissesava_bbu_zustimmung;
  ava_gesprachstyp?: Ava_bbu_beurteilungbesondererbedurfnissesava_gesprachstyp;
  ava_name: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Ava_bbu_beurteilungbesondererbedurfnissesstatecode;
  statuscode?: Ava_bbu_beurteilungbesondererbedurfnissesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Ava_bbu_beurteilungbesondererbedurfnisses extends Ava_bbu_beurteilungbesondererbedurfnissesBase {
  ava_bbu_gesprachsteilnehmerinname?: string;
  ava_bbu_gesprachsteilnehmerinyominame?: string;
  ava_bbu_klientname?: string;
  ava_bbu_klientyominame?: string;
  ava_bbu_standortname?: string;
  ava_bbu_statusname?: string;
  ava_bbu_zustimmungname?: string;
  ava_gesprachstypname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  ava_bbu_gesprachsteilnehmerin?: object;
  _ava_bbu_gesprachsteilnehmerin_value?: string;
  ava_bbu_klient?: object;
  _ava_bbu_klient_value?: string;
  ava_bbu_standort?: object;
  _ava_bbu_standort_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}