// src/generated/models/ContactsModel.ts
export const Contactsava_clientgroup = {
  100000002: 'ARF',
  100000003: 'ARF_Kind_',
  100000004: 'ARM',
  100000007: 'ARM_Kind_',
  100000005: 'Familie',
  100000000: 'UMF',
  100000006: 'UMF_Kind_',
  100000001: 'uUMF'
} as const;
export type Contactsava_clientgroup = keyof typeof Contactsava_clientgroup;
// ...
export type Contactsava_mothertongue = keyof typeof Contactsava_mothertongue;
// ...
export const Contactsgendercode = {
  1: 'M_nnlich',
  2: 'Weiblich',
  192350000: 'Divers',
  192350001: 'Unbekannt',
  100000002: 'Offen',
  100000003: 'Inter'
} as const;
export type Contactsgendercode = keyof typeof Contactsgendercode;
// ...
export interface Contacts extends ContactsBase {
// ...
  _ava_currentsiteid_value?: string;
// ...
}